fixed image threading bug that made unmodified maps look modified after images finished loading fixed minor image bug that introduced rounding error into image size (which also produced extra events during map loading)
git-svn-id: https://tuftsvue.svn.sourceforge.net/svnroot/tuftsvue/trunk@6755 d66e5487-c4fa-4dda-8c8e-ca23240363c0

+++ VUE2/src/tufts/vue/LWComponent.java
@@ -48,7 +48,7 @@ import edu.tufts.vue.preferences.interfaces.VuePreference;
 /**
  * VUE base class for all components to be rendered and edited in the MapViewer.
  *
- * @version $Revision: 1.386 $ / $Date: 2007-11-18 21:49:01 $ / $Author: sfraize $
+ * @version $Revision: 1.387 $ / $Date: 2007-11-19 06:20:27 $ / $Author: sfraize $
  * @author Scott Fraize
  * @license Mozilla
  */
@@ -3641,13 +3641,15 @@ u                    getSlot(c).setFromString((String)value);
             notify(LWKey.Size, old); // todo perf: can we optimize this event out?
     }
 
-    public static Size ConstrainToAspect(float aspect, float w, float h)
+    public static Size ConstrainToAspect(double aspect, double w, double h)
     {
         // Given width & height are MINIMUM size: expand to keep aspect
             
         if (w <= 0) w = 1;
         if (h <= 0) h = 1;
         double tmpAspect = w / h; // aspect we would have if we did not constrain it
+
+        //if (DEBUG.IMAGE) Log.debug("ConstrainToAspect " + tmpAspect);
         // a = w / h
         // w = a*h
         // h = w/a
